+++ SourceCode/e_commerce_app/lib/widgets/hero_carousel_card.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/material.dart';
+
+import '../models/models.dart';
+
+class HeroCarouselCard extends StatelessWidget {
+  //? ************************************************************************
+  //? this will create a list of widgets, and each of this widgets will be a
+  //? container with the background image from the url list copiad above
+
+  final Category category;
+
+  const HeroCarouselCard({Key? key, required this.category}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.symmetric(
+        horizontal: 5.0,
+        vertical: 20,
+      ),
+      child: ClipRRect(
+          borderRadius: const BorderRadius.all(Radius.circular(5.0)),
+          child: Stack(
+            children: <Widget>[
+              Image.network(category.imageUrl,
+                  fit: BoxFit.cover, width: 1000.0),
+              Positioned(
+                bottom: 0.0,
+                left: 0.0,
+                right: 0.0,
+                child: Container(
+                  decoration: const BoxDecoration(
+                    gradient: LinearGradient(
+                      colors: [
+                        Color.fromARGB(200, 0, 0, 0),
+                        Color.fromARGB(0, 0, 0, 0)
+                      ],
+                      begin: Alignment.bottomCenter,
+                      end: Alignment.topCenter,
+                    ),
+                  ),
+                  padding: const EdgeInsets.symmetric(
+                      vertical: 10.0, horizontal: 20.0),
+                  child: Text(
+                    category.name,
+                    style: Theme.of(context)
+                        .textTheme
+                        .headline2!
+                        .copyWith(color: Colors.white),
+                  ),
+                ),
+              ),
+            ],
+          )),
+    );
+  }
+}
