+++ SourceCode/e_commerce_app/lib/widgets/widgets.dart
@@ -1,2 +1,3 @@
 export 'custom_appbar.dart';
 export 'custom_navbar.dart';
+export 'hero_carousel_card.dart';
